fix: minor fixes across services
- fix accel mode handling in motor_service
- fix memory leaks in wifi_sta and wifi_ap
- add logging for LED service button events
- fix telemetry calibration and minor addressing issue

+++ components/app/sensor_service/include/sensor_service.h
@@ -14,15 +14,15 @@ typedef struct
 {
     bool valid;
     float temperature;
-    int32_t pressure;
+    float pressure; 
 } bmp280_data_t;
 
 typedef struct
 {
     bool valid;
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    float x; 
+    float y; 
+    float z; 
 } accel_data_t;
 
 typedef struct
