+++ components/drivers/lsmd6ds3/include/lsmd6ds3.h
@@ -4,4 +4,4 @@
 #include <stdint.h>
 
 esp_err_t lsm6ds3_init(int cs_pin);
-esp_err_t lsm6ds3_read_accel(int16_t *x, int16_t *y, int16_t *z);
+esp_err_t lsm6ds3_read_accel(float *x, float *y, float *z);
